Unbound: animated-material review fixes — colour types index by key frame (`color` held per key frame, `colorLerp` holds the last colour past the last key frame, no past-the-end reads), Lagrange in f64, 16-bit flipbook indices + textureCount, one DISPS open per frame; reader lets scroll layer ranges wrap per SPEC §2 and keeps only the memory-guarding rejections, now stated in SPEC §4.2 (+ §10 clarification, no version bump); reader/exporter share name↔id tables in UnboundSchema.h and the exporter round-trips the whole texture list; scene-state reset moves to OTRPlay_InitScene; Lake Hylia example corrected (old signs scrolled the vanilla direction), env-colour side effect documented; play-tested with a Prelude export

+++ soh/soh/resource/type/scenecommand/SetAnimatedMaterialList.h
@@ -36,7 +36,7 @@ class SetAnimatedMaterialList : public SceneCommand<AnimatedMaterial> {
         AnimatedMatTexCycleParams params;
         std::vector<std::string> texturePaths; // "__OTR__<path>"; the interpreter resolves the name at draw
         std::vector<void*> textureList;        // c_str() of the above
-        std::vector<u8> frames;
+        std::vector<u16> frames;
     };
 
     AnimatedMaterial& AddScroll(u8 segment, u8 pass, u8 type, const ScrollStorage& storage);
